+++ utility/Constants.h
@@ -76,7 +76,7 @@ enum {FILTER_STAGE_3=3};
 enum {FILTER_COEF_COUNT=5};
 extern const double allpass_filter_coefs[FILTER_COEF_COUNT];
 
-extern const long frequency_stop_threshold;
+extern const size_t frequency_stop_threshold;
 
 // Interrupts
